requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
PyJWT==2.3.0
python-decouple==3.6
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
whitenoise==6.0.0
